Don't allow `:` in file names. (#3972)
File paths with a colon in them can refer to an NTFS "alternate data
stream" which does not operate like a normal file and can break many use
cases. They should be avoided (except for the case of the Windows drive
letter, like "C:").

This reverts some of the test changes made in
87586dfb41df57de791923ac6d8285b66607fca2 which removed the `:` from
the virtual entry names. This seems to be a convention in some plugins.
The test now tests that those names don't result in invalid Windows file
paths.

Co-authored-by: Lukas Taegert-Atkinson <lukastaegert@users.noreply.github.com>